fix(admin): 修复商品状态更新中的类型转换问题
- 将 constant.StatusInvalid 转换为 int16 类型以匹配服务层接口要求
- 将 constant.StatusValid 转换为 int16 类型以确保数据类型一致性
- 解决了因类型不匹配导致的商品状态更新失败问题

+++ app/controller/admin_good.go
@@ -156,7 +156,7 @@ func AdminGoodDisable(ctx *gin.Context) {
 	if !ok {
 		return
 	}
-	st := constant.StatusInvalid
+	st := int16(constant.StatusInvalid)
 	if err := service.Good().AdminUpdate(ctx, id, service.AdminUpdateGoodReq{Status: &st}); err != nil {
 		if errors.Is(err, errno.ErrGoodNotFoundOrNoPermission) {
 			reply.ReplyErrWithMessage(ctx, "商品不存在")
@@ -174,7 +174,7 @@ func AdminGoodRestore(ctx *gin.Context) {
 	if !ok {
 		return
 	}
-	st := constant.StatusValid
+	st := int16(constant.StatusValid)
 	if err := service.Good().AdminUpdate(ctx, id, service.AdminUpdateGoodReq{Status: &st}); err != nil {
 		if errors.Is(err, errno.ErrGoodNotFoundOrNoPermission) {
 			reply.ReplyErrWithMessage(ctx, "商品不存在")
